Temp (#5)
* Create catalog.js

* Update index.js

* Create transactions.js

* Create checkout.js

* Update transactions.js

* Update catalog.js

* Update catalog.js

* Update index.js

* Update index.js

* Update index.js

* Update index.js

* Update catalog.js

* Update checkout.js

* Update transactions.js

* Update app.js

* Update catalog.js

* Update index.js

* Update checkout.js

* Update transactions.js

* Update app.js

* Update app.js

* Update index.js

* Update catalog.js

* Update index.js

* Update checkout.js

* Update transactions.js
